+++ projects/angular-sdk/src/i18n/es.ts
@@ -496,7 +496,7 @@ export const locale: I18n = {
             },
             CONFIRMATION_MESSAGE_1: "Tu vuelo ha sido cancelado y tu reembolso está siendo procesado.",
             CONFIRMATION_MESSAGE_2: "En breve se enviará un correo electrónico de confirmación a {{ userEmail }} con detalles sobre cómo usar los créditos de las aerolíneas",
-            REFUND_DETAILS: "{{ amount }} en {{ method }}",
+            REFUND_DETAILS: "{{ amount }}",
             PASSENGER_TYPE: {
                 ADULT: "Adulto",
                 CHILD: "Niño",